create/edit file structure update
